Separate seeding from data generator and fix CRUD
