+++ database/new_data_generator.js
@@ -1,4 +1,3 @@
-const db = require ('./index.js');
 const faker = require ('faker');
 
 const starters = ['the', 'a'];
@@ -321,29 +320,6 @@ const generalInfo = () => {
   return info;
 };
 
-const insertOverviews = () => {
-  for (let i = 0; i < 100; i++) {
-    //General
-    let general = generalInfo ();
-
-    let generalQuery = `INSERT INTO General (restaurant_name, description, telephone, website, chef,
-      avg_rating, num_ratings, style, dress_code, catering, price_range, private_dining,
-      private_url, latitude, longitude, addr, neighborhood, cross_street, parking, public_transport, payment_options, cuisines, top_tags, additional_tags, hours) 
-      VALUES ('${general.restaurant_name}', '${general.description}', '${general.telephone}', '${general.website}', '${general.chef}',
-      '${general.avg_rating}', '${general.num_ratings}', '${general.style}', '${general.dress_code}', '${general.catering}', '${general.price_range}', '${general.private_dining}',
-      '${general.private_url}', '${general.latitude}', '${general.longitude}', '944 Market Street', '${general.neighborhood}', '945 Market Street', '${general.parking}', '${general.public_transport}',
-      '${general.payment_options}', '${general.cuisines}', '${general.top_tags}', '${general.additional_tags}', '${general.hours}');`;
-
-    db.connection.query (generalQuery, (err, result) => {
-      if (err) {
-        console.log('general', err);
-        throw err;
-      } else {
-        // console.log ('general');
-      }
-    });
-
-  }
+module.exports = {
+  generalInfo,
 };
-
-setTimeout (insertOverviews, 60000);
