+++ package.json
@@ -9,7 +9,7 @@
   },
   "scripts": {
     "oldseed": "mysql -u root < ./database/schema.sql && node ./database/data_generator.js",
-    "newseed": "mysql -u root < ./database/newSchema.sql && node ./database/new_data_generator.js",
+    "newseed": "mysql -u root < ./database/newSchema.sql && node ./database/seed_mysql.js",
     "server-dev": "nodemon server/server.js",
     "start": "node server/server.js && node ./database/data_generator.js",
     "react-dev": "webpack -d --watch"
